// src/app/models/mission.model.ts
export class Mission {
  uuid: string;
  type: string;
  title: string;
  alternative_mission: string;
  description: string;
  the_kill_teams: string;
  the_battlefield: string;
  scouting_phase: string;
  deployment: string;
  battle_length: string;
  victory_conditions: string;
  two_player_map: string;
  four_player_map: string;
  special_rule_image: string;
  source: string;
  page: string;

  specialRulesTitle: string;
  specialRules: string;
  specialRules2Title: string;
  specialRules2: string;

  attackerTactic: string;
  defenderTactics: string;

  resources: string;

  requiresCommander: boolean;

  contributor: string;
}